Default install path checked last for op location
If Detectron happens to be installed on the system and someone wants to add their own local installation, they will either add the necessary folders to the python path or the library can be found in the local system path (e.g. the path to the anaconda environment used). With the default install path ("_CMAKE_INSTALL_PREFIX") moved to last position, the system-wide installation won't override the other search paths.