Fix a bug in RPN proposal label generation
Summary:
Negative labels were never set if `len(bg_inds) <= num_bg`. This
condition happens when there aren't as many negative anchor samples compared to the
typical batch size (256). Since detection scales are typically ~800, there are
plenty of negative samples and this bug is likely never hit. In any case, it
is certainly a bug which should be fixed!

Reviewed By: rbgirshick

Differential Revision: D9647703

fbshipit-source-id: fada9f4f2e0a0b5c1619940328c505e73f9a1c7e