Filter proposals not present in roidb
Summary:
It is convenient to run with a set of precomputed proposals for a
larger set (e.g., coco_trainval2014) and then do experiments using annotations
for a smaller subset of the data.

It is possible to do this filtering only if a config flag is set, but it seems
harmless enough to do it always.

Reviewed By: rbgirshick

Differential Revision: D14106953

fbshipit-source-id: f726425ee5b6b22fcb711e3ec479b302ea09e837