replace np.float with float [1]
Summary:
`numpy.float` has been deprecated and removed in numpy-1.20.0. It was an alias to the builtin `float` and should be safe to replace with `float`.

The whole change was mechanical with the following one-liner:
```
fbgr -sl '\bnp\.float\b' | xargs perl -pi -e 's,\bnp\.float\b,float,g'
```

Differential Revision: D46609272

fbshipit-source-id: 3dffc034f3d45a2a06143fca0a93db9b9c4a6feb